refactor(cli): extract playwright-runner + exit-codes modules
Move the Playwright spawn-and-parse logic (runPlaywrightTests,
processPlaywrightResults, the abort/results readers, and the
PlaywrightResult/AbortReason types) out of e2e.ts into
src/cli/utils/playwright-runner.ts. The runner takes a narrow RunGuideOptions
type rather than the whole command options, so it does not depend on the e2e
command surface.

Extract the shared ExitCode contract into src/cli/utils/exit-codes.ts so both
the command and the runner can import it without a cycle back into e2e.ts.
